Improve traverseArray from quadratic to linear
We use a list as an intermediate representation while building up the
result, which has the benefit of O(1) cons, rather than O(n). Then, we
just need to convert this list to an array in the final step, which is
also O(n). So the complexity of traverse is reduced to O(n) from O(n^2).