Added `mapDefault` and `foldMapDefault` to `Data.Traversable`.
These functions implement `map` and `foldMap` in terms of `traverse`,
which is a useful way of reducing boilerplate in instances of `Functor`
and `Foldable` when you have an instance of `Traversable`.

I didn't want to add new dependencies on `purescript-const` and
`purescript-identity` (especially since they'd be circular dependencies),
so I wrote short (private) implementations of the `Const` and `Identity`
applicatives in the `Traversable` file directly.