Add -WithIndex classes
This adds `FunctorWithIndex`, `FoldableWithIndex` and
`TraversableWithIndex` and some combinators. Implementations are mostly
copied from the versions without index with minor adaptions. Similarly
for the documentation.

I think the corresponding haskell package does not state all laws for
these type classes, so there are TODO's in the doc strings of the three
classes.

`StateL` and `StateR` in TraversableWithIndex.purs are exact copies of
the versions in Traversable.purs because they are not exported there.
Maybe it would be better to merge TraversableWithIndex.purs into
Traversable.purs and similarly for Foldable. The Foldable and Traverse
instances of Array could then be along the lines of
```purescript
traverse f = itraverse (const f)`
```
so that the foreign js imports for the unindexed version could be
removed.

`FunctorWithIndex`'s `imap` collides with the `imap` of
purescript-invariant. Haskell calls the latter `invmap`. I don't know
whether that's a problem.

This does not yet include the `-1WithIndex` classes, i.e.
generalizations of `Foldable1` and `Traversable1` in `Semigroup/`. I
guess it would make sense to provide them here as well.