Refactor test to detect foldMapDefaultL bug
This refactors the test for the Foldable instance of array and the
default implementations of Foldable's functions.

In a way, it does not really make sense to test the default
implementations of Foldable's functions, because they are part of the
rules of the type class. Up to execution speed and memory usage, every
valid instance has to satisfy
```purescript
foldMap = foldMapDefaultL
foldMap = foldMapDefaultR
foldl = foldlDefault
foldr = foldrDefault
```
and, for any reasonable instance that is also a functor,
```purescript
foldMap f = fold <<< map f
```
Testing these properties for some instances that we think are valid
should reveal errors in the specification.

I've refactored the test so that the equations above can be tested via
`foldMap` on any Monoid m, in particular on non-commutative monoids such
as `Array a`, which detects the `foldMapDefaultL` bug.